final commits

+++ src/main/java/com/javarush/domain/repository/CityRepository.java
@@ -3,8 +3,6 @@ package com.javarush.domain.repository;
 import com.javarush.domain.entity.City;
 import jakarta.persistence.EntityManager;
 
-import java.util.List;
-
 public class CityRepository extends RepositoryBase<Integer, City> {
 
     public CityRepository(EntityManager entityManager) {
@@ -13,7 +11,7 @@ public class CityRepository extends RepositoryBase<Integer, City> {
 
     public long getTotalCount() {
         var query = getEntityManager().createQuery("select count(c) from City c", Long.class);
-        return (long) query.getSingleResult();
+        return query.getSingleResult();
     }
 
 }
